sortwaterai-bot/requirements.txt
aiogram==2.25.1
python-dotenv==0.21.0
psycopg2-binary==2.9.9
